turring-backend-mvp/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic-settings==2.1.0
redis==5.0.8
openai==2.8.1
pyjwt==2.8.0
bcrypt==5.0.0
python-multipart==0.0.9
